Made the cmakelists accept the new header structure. Renamed NovelStepTimer to StepTimer.
