+++ include/NovelRT.h
@@ -1,37 +1,36 @@
-#ifndef NOVELRT_H
-#define NOVELRT_H
+#ifndef RT_H
+#define RT_H
 
-#include "CameraBlock.h"
-#include "CameraViewChangedEventArgs.h"
-#include "GeoBounds.h"
-#include "GeoMatrix4.h"
-#include "GeoVector.h"
-#include "GraphicsCharacterRenderData.h"
-#include "KeyCode.h"
-#include "KeyState.h"
-#include "Lazy.h"
-#include "NovelAudioService.h"
-#include "NovelBasicFillRect.h"
-#include "NovelBasicInteractionRect.h"
-#include "NovelCamera.h"
-#include "NovelCommonArgs.h"
-#include "NovelDebugService.h"
-#include "NovelImageRect.h"
-#include "NovelInteractionObject.h"
-#include "NovelInteractionService.h"
-#include "NovelLayeringService.h"
-#include "NovelLoggingService.h"
-#include "NovelObjectSortComparison.h"
-#include "NovelRenderingService.h"
-#include "NovelRenderObject.h"
-#include "NovelRunner.h"
-#include "NovelRunner_C.h"
-#include "NovelStepTimer.h"
-#include "NovelTextRect.h"
-#include "NovelUtilities.h"
-#include "NovelWindowingService.h"
-#include "NovelWorldObject.h"
-#include "RGBAConfig.h"
-#include "ShaderProgram.h"
-#include "wrapper.hpp"
-#endif //!NOVELRT_H
+#include "NovelRT/Graphics/CameraBlock.h"
+#include "NovelRT/Graphics/CameraViewChangedEventArgs.h"
+#include "NovelRT/Maths/GeoBounds.h"
+#include "NovelRT/Maths/GeoMatrix4.h"
+#include "NovelRT/Maths/GeoVector.h"
+#include "NovelRT/Graphics/GraphicsCharacterRenderData.h"
+#include "NovelRT/Input/KeyCode.h"
+#include "NovelRT/Input/KeyState.h"
+#include "NovelRT/Utilities/Lazy.h"
+#include "NovelRT/Audio/AudioService.h"
+#include "NovelRT/Graphics/BasicFillRect.h"
+#include "NovelRT/Input/BasicInteractionRect.h"
+#include "NovelRT/Graphics/Camera.h"
+#include "NovelRT/Utilities/CommonArgs.h"
+#include "NovelRT/DebugService.h"
+#include "NovelRT/Graphics/ImageRect.h"
+#include "NovelRT/Input/InteractionObject.h"
+#include "NovelRT/Input/InteractionService.h"
+#include "NovelRT/LayeringService.h"
+#include "NovelRT/LoggingService.h"
+#include "NovelRT/TransformSortComparison.h"
+#include "NovelRT/Graphics/RenderingService.h"
+#include "NovelRT/Graphics/RenderObject.h"
+#include "NovelRT/NovelRT-C/NovelRunner.h"
+#include "NovelRT/NovelRT-C/NovelRunner_C.h"
+#include "NovelRT/Timing/StepTimer.h"
+#include "NovelRT/Graphics/TextRect.h"
+#include "NovelRT/Utilities/Misc.h"
+#include "NovelRT/Windowing/WindowingService.h"
+#include "NovelRT/WorldObject.h"
+#include "NovelRT/Graphics/RGBAConfig.h"
+#include "NovelRT/Graphics/ShaderProgram.h"
+#endif //!RT_H
